pyproject.toml
[project]
name = "switch"
version = "0.4.0"
description = "A program to switch between projects"
requires-python = ">=3.11"
dependencies = [
    "argparse",
    "python-ulid",
    "toml",
    "typing-extensions",
    "inquirer",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
switch = "switch.main:main"
